added implicit getters to VaultController interface and tested
Former-commit-id: 3ed839118c9a29071a5d2920c6c1c769f273804c [formerly 00a3a51310374499aa9b4e20eb9263afafdd40ba] [formerly 4e0bd9d6aa0a20cd964e9352914bde2672d3c6e0 [formerly 3f818cb55399684f5db956f9a9e7d17bcd323669]]
Former-commit-id: 73d3ff5c44885ba03efdf14543d6ca948c3ee800
Former-commit-id: 572b29f4de85ea35aa8b5755d6188e892fbd4972

+++ contracts/lending/VaultController.sol
@@ -102,13 +102,13 @@ contract VaultController is
     return _vaultId_vaultAddress[id];
   }
   
-  /**
+  
   /// @notice get total base liability of all vaults
   /// @return total base liability
-  function totalBaseLiability() external view override returns (uint256) {
+  function totalBaseLiability() external view override returns (uint192) {
     return _totalBaseLiability;
   }
-   */
+  
   
 
   /// @notice create a new vault
